feat: Observatory update — metrics, data, Klein physics sim
observatory.py:
  - Updated metrics: 50 Lean, 11 Rust, 103 tests, 0 sorry
  - Updated footer counts

Resonance Analysis:
  - Generated zeta_divergence_analysis.csv (5,050 rows)
  - Computed from 100K real zeta zeros + 100 primes
  - LaRue antenna decomposition (Base - Modulator - Drain)
  - Page now loads with full data

Klein Universe (6_Klein_Universe.py):
  - Complete rewrite: N-body Protoreal physics simulation
  - Particles undergo Klein multiplication (non-commutative!)
  - Sowing (funct) + Fibonacci consolidation
  - Parity damping → Hodge class (ω = ι)
  - Convergence to a=1 (Duality Theorem)
  - 3D scatter with holochain trails, SR coloring
  - Real-time convergence + energy metrics
  - Computed γ₀ as noise scale (not hardcoded)

ProtorealEngine.py:
  - Level (λ) now float, not int (fixes Klein mul NaN)

+++ SpectralObservatory/core/ProtorealEngine.py
@@ -10,7 +10,7 @@ class ProtorealElement:
         self.b = float(b) # Thrust (ω)
         self.c = float(c) # Anchor (ι)
         self.e = float(e) # Noise (ε)
-        self.l = int(l)   # Level (λ)
+        self.l = float(l)  # Level (λ)
 
     def __str__(self):
         return f"{{a: {self.a:.4f}, ω: {self.b:.4f}, ι: {self.c:.4f}, ε: {self.e:.4f}, λ: {self.l}}}"
